Implement workaround to enable multi-app (#145)
Aviary currently combines all `LLMDeployments` and the `Router`
deployment into a single Serve app. When a request arrives, the `Router`
determines which `LLMDeployment` should process it and it sends the
request to that deployment over a Serve handle. However, due to a
limitation in Ray and Ray Serve, deployments called from other
deployments using a Serve Handle cannot stream their responses. This
means the current single-app architecture cannot support streaming.

This change enables streaming by splitting the `Router` and each
`LLMDeployment` into its own Serve app. To work around the limitation,
whenever the `Router` app receives an HTTP request, it creates a new
HTTP request over `localhost` to a streaming `LLMDeployment`, receives
the response stream, and sends a `StreamingResponse` containing that
stream back to the original client. This workaround is meant to be
temporary and should be replaced with a longer-term solution once the
Ray Serve limitation is fixed.

---------

Signed-off-by: Max Pumperla <max.pumperla@googlemail.com>
Signed-off-by: Shreyas Krishnaswamy <shrekris@anyscale.com>
Signed-off-by: Antoni Baum <antoni.baum@protonmail.com>
Co-authored-by: Max Pumperla <max.pumperla@googlemail.com>
Co-authored-by: Kyle Huang <kyle@anyscale.com>
Co-authored-by: M Waleed Kadous <mwk@anyscale.com>
Co-authored-by: sijie <sijie@anyscale.com>
Co-authored-by: Antoni Baum <antoni.baum@protonmail.com>